Begin work on trend calibration script
The trend calibration script compares calculated polling trends
for past election periods to the eventual result and determines
the adjustments that need to be made to create a calibrated
distribution from a basic poll trend, projected to any time
in the present or future. The work in this commit gets the
code to the point where it can load polling data and store it.
Separately, the 2022-fed file has been updated with the latest
poll trends, with only slight impacts on the simulations.