Further changes to minor party fp analysis.
Specifically, dummy values (0 or 1) are now output for
any parties and parameters where the actual calculation of that
parameter would be inappropriate. For example, the "recontest
rate" for Greens is set to be equal to 1, and the "swing
coefficient" for Independents is set to be equal to zero. Also, for
primary vote buckets where no sophomores exist historically,
the sophomore coefficient is set to be equal to that for the next
smallest bucket that has at least one sophomore. This allows
for smoothing and also use in cases where the winner does
have a primary vote that low.