Various fixes and updates to poll files.
- Include UAP figure for Essential and update files
related to that
- Fix sum-to-zero house effect constraint applying
to some parties it shouldn't and made changes to
make this less likely in future
- Add an old Nielsen poll to 2006vic
- Update preference flows for 2022fed.
Special note: for now UAP preference flows are
being given as the average of 2013/2019
elections. This is because while the 2019 UAP
campaigned along specifically anti-ALP lines,
2013 PUP was more generally anti-majors and
the present campaign is more in line with that,
so using last-election preferences may
underestimate the ALP's preference flow
somewhat.