Add new region base behaviour.
For projecting region behaviour without region
polls (or where such polls are not useful). Note that
ALP chance/vote increases a bit in the 2022fed
save file, this is mainly due to re-running trend files
and not a significant effect of the regional
adjustments (which do however shift the balance
of seats between regions a little).