Handle simulation modes and dates.
Simulation modes are e.g. regular forecast, live forecast,
nowcast. This is now sent to the server which will store
each separately. Dates are now sent to the server converted
to UTC. Also prevented "Latest Forecast" from being
uploaded to server as this is expected to cause mistakes
later on.