Normalisation for tpp-first trends using fundamentals for minor parties.
This caps the chances of minor parties a bit at the ends,
especially that of two or more doing well at the same time, which
should especially lower the chances of ONP vs. GRN tcp scenarios
and similar oddities, and is consistent with observations that
minor parties tend to crowd each other out rather than simply
taking separate shares of major party votes.