Adjust IND->ALP prefs according to historical data.
Previous changes assumed an IND->ALP preference rate that
was far too high when the IND fp was very high. This change
brings the IND vote in line with historical data.
Additionally, adjusted ind vote for confirmed IND candidates
(previously only the emergence rate was affected), and made
GRN->INDvsLNP an explicit value (of 85%) as it was somewhat
underestimated and we have evidence it's around this level.