Adjustments for live 2023nsw + fix wonky prob curve issue
-"Wonky prob curve issue" refers to the
fact that the uncertainty would at a certain
point on election night erroneously begin to
increase rather than decrease as more votes
come in. This was a result of a system designed
to handle PPVCs and dec votes but became
overly complex and created unexpected
behaviours that were difficult to fix. It has
for this election been replaced with a much
simpler system to address the issue of
potential reporting bias across the entire
election.